libutil/unlink_recursive: factor out of cleanup.c
Move unlink_recursive() to its own C module so that it can be
tested and used independently.
