+++ src/common/libutil/cleanup.c
@@ -28,22 +28,15 @@
 
 #include "cleanup.h"
 #include "xzmalloc.h"
+#include "unlink_recursive.h"
 
 #include <stdio.h>
 #include <stdlib.h>
-#include <sys/stat.h>
-#include <sys/types.h>
-#include <fcntl.h>
-#include <dirent.h>
 #include <libgen.h>
 #include <unistd.h>
 #include <pthread.h>
 
 #include <czmq.h>
-
-#ifndef O_PATH
-#define O_PATH 0 /* O_PATH is too new for CentOS 6 - see issue #646 */
-#endif
 
 static pthread_mutex_t mutex = PTHREAD_MUTEX_INITIALIZER;
 
@@ -51,40 +44,6 @@ struct cleaner {
     cleaner_fun_f * fun;
     void * arg;
 };
-
-static void unlinkat_recursive (int fd)
-{
-    DIR *dir;
-    struct dirent *d, entry;
-    struct stat sb;
-    int cfd;
-
-    if ((dir = fdopendir (fd))) {
-        while (readdir_r (dir, &entry, &d) == 0 && d != NULL) {
-            if (!strcmp (d->d_name, ".") || !strcmp (d->d_name, ".."))
-                continue;
-            if (fstatat (fd, d->d_name, &sb, AT_SYMLINK_NOFOLLOW) < 0)
-                continue;
-            if (S_ISDIR (sb.st_mode)) {
-                if ((cfd = openat (fd, d->d_name, O_PATH | O_DIRECTORY)) < 0)
-                    continue;
-                unlinkat_recursive (cfd);
-                (void)unlinkat (fd, d->d_name, AT_REMOVEDIR);
-            } else
-                (void)unlinkat (fd, d->d_name, 0);
-        }
-        closedir (dir);
-    }
-}
-
-static void unlink_recursive (const char *dirpath)
-{
-    DIR *dir = opendir (dirpath);
-    if (dir) {
-        unlinkat_recursive (dirfd (dir));
-        (void)rmdir (dirpath);
-    }
-}
 
 void cleanup_directory_recursive (const struct cleaner *c)
 {
