+++ src/common/libutil/unlink_recursive.h
@@ -1,0 +1,6 @@
+void unlink_recursive (const char *dirpath);
+
+/*
+ * vi:tabstop=4 shiftwidth=4 expandtab
+ */
+
